==> scripts/00_clean_ais_file.Rmd <==
---
title: "Dominica: Process Commercial Vessel Tracks"
output: html_document
date: "2023-05-01"
---

# Introduction

This script loads and cleans the database of vessel positions around Dominica between 2012 - 2019 shared with us by Shane Gero. 

```{r setup, include=FALSE}
# Libraries
library(lubridate)

# Load common helper script
source(here::here("common.R"))

# # Load reserve boundaries
load(here::here("results", "reserve_boundaries.Rdata"))

# Load raw ais data
ais_dat <- read_csv(here::here("data", "ais", "raw", "shiny_vessels.csv")) %>%
  janitor::clean_names()
```

# Clean dates and column names

```{r}
# Process the data -----------------------------------------------------------
ais_dat_clean <- ais_dat %>%
  as.data.table() %>%
  dplyr::select(mmsi, vessel_type, category, datetime = timestamp, lat, lon, speed) %>%
  mutate(lat = as.numeric(lat),
         lon = as.numeric(lon),
         year = lubridate::year(datetime),
         month = lubridate::month(datetime))

# Export the file ------------------------------------------------------------
write_csv(ais_dat_clean, file = here("data", "ais", "clean", paste0("Dominica_AIS_2012_2019.csv")))
```

# Segment

The AIS positions provided to us by Shane don't appear to have been grouped into any sort of continuous segments or trips, so let's do this first.

```{r}
# Create segments  ------------------------------------------------------------
ais_dat_seg <- ais_dat_clean %>%
  group_by(mmsi) %>%
  mutate(hours = difftime(datetime, lag(datetime), units = "hours")) %>% # time difference in hours
  ungroup() %>%
  mutate(start_new_seg = case_when(is.na(hours) ~ T,
                                   hours > 24 ~ T), # If more than 24 hours has elapsed, start new segment
         seg_id = ifelse(start_new_seg, paste0(mmsi, "_", datetime), NA),
         seg_id = fill.na(seg_id)) %>%
  dplyr::select(-start_new_seg)

ais_dat_seg$hours[is.na(ais_dat_seg$hours)] <- 0
```

# Calculate distances between points

Next we want to get the distances between points so we can determine changes in trip distance. Using the tidyverse approach below, this step takes ~45 minutes to run on my computer. 

```{r}
# Make coords spatial ----------------------------------------------------------
ais_dat_sf <- ais_dat_seg %>%
  st_as_sf(coords = c("lon", "lat"),
           crs = 4326) %>%
  sf::st_transform(crs = st_crs(prj_moll)) # convert points to Mollewide

# Calculate distances ----------------------------------------------------------
ais_dat_dist <- ais_dat_sf %>%
  group_by(mmsi, seg_id) %>%
  mutate(dist_from_last = st_distance(lag(geometry), geometry, by_element = T),
         dist_to_next = st_distance(geometry, lead(geometry), by_element = T)) %>%
  ungroup() %>%
  mutate(dist_from_last = ifelse(!is.na(dist_from_last), dist_from_last, 0),
         dist_to_next = ifelse(!is.na(dist_to_next), dist_to_next, 0),
         dist = 0.5*dist_from_last + 0.5*dist_to_next)

# Export the file ------------------------------------------------------------
save(ais_dat_dist, file = here("data", "ais", "clean", "Dominica_AIS_2012_2019_processed.Rdata"))
```

# Create segment summary table

Let's create a summary table of all segments in our data. 

```{r}
# Create summary table ------------------------------------------------------------
seg_summary_table <- ais_dat_dist %>%
  st_transform(crs = 4326) %>%
  group_by(seg_id) %>%
  summarize(n_pos = n(),
            n_pos_moving = length(hours[speed >= 0.1]),
            hours = sum(hours, na.rm = T),
            active_hours = sum(hours[speed >= 0.1], na.rm = T),
            first_timestamp = min(datetime),
            last_timestamp = max(datetime),
            min_lon = min(st_coordinates(geometry)[,1]),
            max_lon = max(st_coordinates(geometry)[,1]),
            min_lat = min(st_coordinates(geometry)[,2]),
            max_lon = max(st_coordinates(geometry)[,2]),
            average_speed = mean(speed, na.rm = T),
            year_start = min(year),
            year_end = max(year)) %>%
  st_drop_geometry()

# Export the file ------------------------------------------------------------
save(seg_summary_table, file = here("data", "ais", "clean", "Dominica_AIS_segment_summary.Rdata"))
```


## Visualize 

```{r}
all_points_by_category_2019 <- ais_dat_dist %>%
  dplyr::filter(year == 2019) %>%
  group_by(seg_id, category) %>%
  summarize(n_pos = n(),
            average_speed = mean(speed, na.rm = T))

# Plot by category. All over the place.
all_tracks_by_category_2019_plot <- ggplot()+
  geom_sf(data = dominica_eez_sf_moll, aes(group = mrgid), color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = all_points_by_category_2019, size = 0.05, fill = "black")+
  geom_sf(data = reserve_boundaries_sf, aes(fill = scenario), alpha = 0.3)+
  coord_sf(xlim = plot_bbox$xlim, ylim = plot_bbox$ylim)+
  theme_basic()+
  facet_wrap(~category)+
  guides(fill = "none")

ggsave(here("results", "ais_tracks_by_category_2019.png"), all_tracks_by_category_2019_plot)
```

# Extract tracks for the high-speed ferries

```{r}
ferry_ais_positions <- ais_dat_dist %>%
  dplyr::filter(category == "high_speed_ferry")

# Export the file ------------------------------------------------------------
save(ferry_ais_positions, file = here("data", "ais", "clean", "Dominica_AIS_ferry_only.Rdata"))
```
