==> scripts/02_whale_sightings.Rmd <==
---
title: "Dominica: Sperm Whale Sightings"
output: html_document
date: "2023-05-01"
---

```{r setup, include=FALSE}
# Load common helper script
source(here::here("common.R"))

# Define nautical miles in m
nm <- 1852

# Load reserve boundaries
load(file.path(prj_cachalote_dir, "data", "processed", "reserve_boundaries.Rdata"))

# Load Dominica boundaries 
load(file.path(prj_cachalote_dir, "data", "processed", "dominica_boundaries.Rdata"))

# Load shipping lane and buffer zone boundaries
load(file.path(prj_cachalote_dir, "data", "processed", "buffer_and_shipping_lane_boundaries.Rdata"))
```

# Introduction

This script loads the database of sperm whale sightings around Dominica from 2005 - 2018 shared with us by Shane Gero and assessed the proportion of total sightings each year that fall within the boundaries of the proposed reserve(s). 

# Explore data

```{r}
# Load sperm whale sightings
whales <- read_csv(file.path(prj_cachalote_dir, "data", "raw", "sightings.csv")) %>%
  rowwise() %>%
  mutate(datetime_rev = as.POSIXlt(datetime, format = "%m/%d/%Y %H:%M"),
         year = year(datetime_rev)) %>%
  st_as_sf(coords = c("lon", "lat"),
           crs = 4326,
           remove = F) %>%
  rename(obs = 1) %>%
  sf::st_transform(crs = st_crs(prj_moll)) %>% # convert points to Mollewide
  rowwise() %>%
  mutate(x = st_coordinates(geometry)[1],
         y = st_coordinates(geometry)[2])

# Filter for years of interest
whales_recent <- whales %>%
  dplyr::filter(year > 2013)
```

```{r}
# Add scenario name
reserve_boundaries_sf <- reserve_boundaries_sf %>%
  mutate(scenario_name = paste0("Scenario ", scenario))

# Plot 
whales_and_reserves_plot <- ggplot()+
  geom_sf(data = whales_recent, size = 0.02, fill = "black")+
  geom_sf(data = dominica_sf_moll, aes(group = name), color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = reserve_boundaries_sf, aes(color = rev(scenario_name)), alpha = 1, fill = NA, size = 0.5)+
  coord_sf(xlim = plot_bbox$xlim, ylim = plot_bbox$ylim)+
  facet_wrap(~scenario_name, ncol = 1)+
  theme_basic()+
  guides(color = "none")+
  labs(title = "")

save_plots(file_name = file.path(prj_cachalote_dir, "figures", "potential_reserve_boundaries_and_whale_sightings_map.png"), 
           plot_name = whales_and_reserves_plot,
           height_in = 9,
           width_in = 9)

whales_and_reserves_plot
```

```{r}
# Plot again by year
whales_and_reserves_plot_by_year <- ggplot()+
  geom_sf(data = dominica_sf_moll, color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = whales_recent, size = 0.02, fill = "black")+
  geom_sf(data = reserve_boundaries_sf, aes(color = rev(scenario_name)), alpha = 1, fill = NA, size = 0.5)+
  coord_sf(xlim = plot_bbox$xlim, ylim = plot_bbox$ylim)+
  facet_grid(scenario_name ~ year)+
  theme_basic()+
  guides(color = "none")+
  labs(title = "Potential reserve boundaries overlaid on sperm whale sightings (2014 - 2018)")

save_plots(file_name = file.path(prj_cachalote_dir, "figures", "potential_reserve_boundries_and_whale_sightings_by_year_map.png"), 
           plot_name = whales_and_reserves_plot_by_year,
           height_in = 9,
           width_in = 9)

whales_and_reserves_plot_by_year
```

```{r}
whale_labels <- c(1,3,10,30,100,300)
whale_breaks <- log10(whale_labels)

# Density plot
whales_density <- ggplot() +
  geom_bin2d(data = whales_recent, aes(x = x, y = y, fill = log10(..density..)), bins = 100) +
  geom_sf(data = dominica_sf_moll, color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = reserve_boundaries_sf, aes(color = rev(scenario_name)), alpha = 1, fill = NA, size = 0.5)+
  scale_fill_viridis_c(option = "C",
                       labels = whale_labels, 
                       name = "Sightings (count)") +
  coord_sf(xlim = plot_bbox$xlim, ylim = plot_bbox$ylim)+
  theme_basic()+
  guides(color = "none")

save_plots(file_name = file.path(prj_cachalote_dir, "figures", "whale_sightings_density_map.png"), 
           plot_name = whales_density,
           height_in = 9,
           width_in = 9)

whales_density
```

```{r}
dominica_sf_latlon <- dominica_sf_moll %>%
  st_transform(crs = 4326)

reserve_boundaries_latlon <- reserve_boundaries_sf %>%
  st_transform(crs = 4326)

# Density plot
whales_density_lat_lon <- ggplot() +
  geom_bin2d(data = whales_recent %>% st_drop_geometry(), aes(x = lon, y = lat, fill = log10(..density..)), bins = 80) +
  geom_sf(data = dominica_sf_latlon, color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = reserve_boundaries_latlon, aes(color = rev(scenario_name)), alpha = 1, fill = NA, size = 0.5)+
  scale_fill_viridis_c(option = "C",
                       #labels = whale_labels,
                       #breaks = whale_breaks,
                       name = "Sightings (count)") +
  coord_sf(xlim = c(-62,-61), ylim = c(15,15.8))+
  theme_basic()+
  guides(color = "none")

save_plots(file_name = file.path(prj_cachalote_dir, "figures", "whale_sightings_density_map_latlon.png"), 
           plot_name = whales_density_lat_lon,
           height_in = 9,
           width_in = 9)

whales_density_lat_lon
```

# Results - Version 1

## Find whale sightings inside the reserve area(s)

```{r}
# Determine which reserve(s) each observation would fall within
where_be_whales <- st_join(whales_recent, reserve_boundaries_sf)

# Get sightings inside reserve
reserve_whales <- where_be_whales %>%
  dplyr::filter(!is.na(scenario))

# Get all other sightings
outside_whales <- where_be_whales %>%
  dplyr::filter(is.na(scenario))
```

## Summarize by Reserve Design and Year

```{r}
# Get totals by year
totals_by_year <- whales_recent %>%
  st_drop_geometry() %>%
  group_by(year) %>%
  summarize(n_sightings_total = n_distinct(obs),
            n_adults_total = sum(adults, na.rm = T),
            n_calves_total = sum(calves, na.rm = T),
            n_whales_total = sum(total, na.rm = T))
  
# Summarize by year and reserve design
summary_by_year <- reserve_whales %>%
  st_drop_geometry() %>%
  group_by(year, scenario, scenario_name, area_km2) %>%
  summarize(n_sightings = n_distinct(obs),
            n_adults = sum(adults, na.rm = T),
            n_calves = sum(calves, na.rm = T),
            n_whales = sum(total, na.rm = T)) %>%
  ungroup() %>%
  left_join(totals_by_year, by = "year") %>%
  mutate(prop_sightings = n_sightings/n_sightings_total,
         prop_adults = n_adults/n_adults_total,
         prop_calves = n_calves/n_calves_total,
         prop_whales = n_whales/n_whales_total) %>%
  dplyr::select(year, scenario, scenario_name, area_km2, contains("prop")) %>%
  gather("obs_type", "value", contains("prop")) %>%
  mutate(obs_type = str_replace(obs_type, "prop_", ""))
```

## Output Table 

```{r}
out_table <- summary_by_year %>%
  dplyr::filter(obs_type == "sightings") %>%
  distinct(scenario, year, value) %>%
  mutate(value = round(value*100, 2)) %>%
  group_by(scenario) %>%
  mutate(avg = mean(value)) %>%
  spread(year, value) %>%
  relocate(avg, .after = `2018`)

head(out_table)
```

## Visualize

```{r}
# Plot reserve area versus sightings
tradeoff_plot <- summary_by_year %>%
  ggplot()+
  aes(x = area_km2, y = value, color = obs_type)+
  geom_point(aes(shape = scenario_name), size = 2)+
  geom_line()+
  theme_basic()+
  facet_wrap(~ year, ncol = 5)+
  labs(x = "Reserve area (km2)", y = "Prop. of observations", title = "Proportion of sperm whale sightings within potential reserves by area (2014 - 2018)")+
  guides(color = guide_legend(title = "Observation type"),
         shape = guide_legend(title = "", reverse = F))
  
save_plots(file_name = file.path(prj_cachalote_dir, "figures", "prop_whale_sightings_by_reserve_and_year.png"), 
           plot_name = tradeoff_plot,
           height_in = 4,
           width_in = 9)

tradeoff_plot
```

## Summarize by Other Areas

```{r}
# Determine which other areas each observation would fall within
where_be_whales_outside <- st_join(outside_whales, other_areas_sf)

# Summarize by year and reserve design
summary_by_year_outside <- where_be_whales_outside %>%
  mutate(name = ifelse(is.na(name), "Other areas", name)) %>%
  st_drop_geometry() %>%
  group_by(year, name) %>%
  summarize(n_sightings = n_distinct(obs),
            n_adults = sum(adults, na.rm = T),
            n_calves = sum(calves, na.rm = T),
            n_whales = sum(total, na.rm = T)) %>%
  ungroup() %>%
  left_join(totals_by_year, by = "year") %>%
  mutate(prop_sightings = n_sightings/n_sightings_total,
         prop_adults = n_adults/n_adults_total,
         prop_calves = n_calves/n_calves_total,
         prop_whales = n_whales/n_whales_total) %>%
  dplyr::select(year, name, contains("prop")) %>%
  gather("obs_type", "value", contains("prop")) %>%
  mutate(obs_type = str_replace(obs_type, "prop_", "")) 
```

```{r}
# Make summary table
other_table <- summary_by_year_outside %>%
  dplyr::filter(obs_type == "sightings") %>%
  distinct(name, year, value) %>%
  mutate(value = round(value*100, 2)) %>%
  group_by(name) %>%
  mutate(avg = round(mean(value), 2)) %>%
  spread(year, value) %>%
  relocate(avg, .after = `2018`)

head(other_table)
```

```{r}
# Rename 
shipping_lane_sf <- other_areas_sf %>%
 dplyr::filter(name != "Buffer zone - 2nm") %>%
  mutate(scenario = as.numeric(str_replace(name, "Shipping lane - scenario ", "")),
         scenario_name = paste0("Scenario ", scenario))

# Plot 
whales_and_other_areas_plot <- ggplot()+
  geom_sf(data = where_be_whales_outside, size = 0.02, fill = "black")+
  geom_sf(data = dominica_sf_moll, color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = other_areas_sf %>% dplyr::filter(name == "Buffer zone - 2nm"), alpha = 1, color = "orange", fill = NA, size = 0.5)+
  geom_sf(data = shipping_lane_sf, aes(color = fct_rev(scenario_name)), alpha = 1, fill = NA, size = 0.5)+
  coord_sf(xlim = plot_bbox$xlim, ylim = plot_bbox$ylim)+
  theme_basic()+
  guides(color = "none")+
  labs(title = "")

save_plots(file_name = file.path(prj_cachalote_dir, "figures", "other_areas_and_whale_sightings_map.png"), 
           plot_name = whales_and_other_areas_plot,
           height_in = 5,
           width_in = 5)

whales_and_other_areas_plot
```

# Explore results excluding Roseau hotspot 

There's a major concentration of sightings right by Roseau... Like an order of magnitude greater than most other areas. This seems suspicious. 

```{r}
# Bin sightings into lat/lon bins
whales_recent_binned <- whales_recent %>%
  mutate(lat_bin = (floor(lat/0.01)*0.01) + 0.005,
         lon_bin = (floor(lon/0.01)*0.01) + 0.005) %>%
  group_by(lon_bin, lat_bin) %>%
  summarize(count = n())

# Which bins have more than 200 sightings
potential_bad_whales <- whales_recent_binned %>%
  dplyr::filter(count > 200) %>%
  distinct(lat_bin, lon_bin)

# Coordinates to filter by 
lon_lim <- c(potential_bad_whales$lon_bin-0.005, potential_bad_whales$lon_bin+0.005)
lat_lim <- c(potential_bad_whales$lat_bin-0.005, potential_bad_whales$lat_bin+0.005)

# Now let's get those bad entries
bad_whales <- whales_recent %>%
  dplyr::filter(lat >= lat_lim[1] & lat <= lat_lim[2] & lon >= lon_lim[1] & lon <= lon_lim[2])

# Plot to check that we've got the right ones - Yep
bad_whales_plot <- ggplot()+
  geom_sf(data = bad_whales, size = 0.02, fill = "black")+
  geom_sf(data = dominica_sf_moll, aes(group = name), color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = reserve_boundaries_sf, aes(color = rev(scenario_name)), alpha = 1, fill = NA, size = 0.5)+
  coord_sf(xlim = plot_bbox$xlim, ylim = plot_bbox$ylim)+
  theme_basic()+
  guides(color = "none")+
  labs(title = "")

bad_whales_plot
```

# ------------------------------------------------------------------------------

Now let's recreate our results if we remove those whales... 

# Results - Version 2

This version has the whales right off of Roseau removed

## Find whale sightings inside the reserve area(s)

```{r}
# Good whales
good_whales_recent <- whales_recent %>%
  dplyr::filter(!obs %in% bad_whales$obs)

# Determine which reserve(s) each observation would fall within
good_where_be_whales <- st_join(good_whales_recent, reserve_boundaries_sf)

# Get sightings inside reserve
good_reserve_whales <- good_where_be_whales %>%
  dplyr::filter(!is.na(scenario))

# Get all other sightings
good_outside_whales <- good_where_be_whales %>%
  dplyr::filter(is.na(scenario))
```

## Summarize by Reserve Design and Year

```{r}
# Get totals by year
good_totals_by_year <- good_whales_recent %>%
  st_drop_geometry() %>%
  group_by(year) %>%
  summarize(n_sightings_total = n_distinct(obs),
            n_adults_total = sum(adults, na.rm = T),
            n_calves_total = sum(calves, na.rm = T),
            n_whales_total = sum(total, na.rm = T))
  
# Summarize by year and reserve design
good_summary_by_year <- good_reserve_whales %>%
  st_drop_geometry() %>%
  group_by(year, scenario, scenario_name, area_km2) %>%
  summarize(n_sightings = n_distinct(obs),
            n_adults = sum(adults, na.rm = T),
            n_calves = sum(calves, na.rm = T),
            n_whales = sum(total, na.rm = T)) %>%
  ungroup() %>%
  left_join(good_totals_by_year, by = "year") %>%
  mutate(prop_sightings = n_sightings/n_sightings_total,
         prop_adults = n_adults/n_adults_total,
         prop_calves = n_calves/n_calves_total,
         prop_whales = n_whales/n_whales_total) %>%
  dplyr::select(year, scenario, scenario_name, area_km2, contains("prop")) %>%
  gather("obs_type", "value", contains("prop")) %>%
  mutate(obs_type = str_replace(obs_type, "prop_", ""))
```

## Output Table 

```{r}
good_out_table <- good_summary_by_year %>%
  dplyr::filter(obs_type == "sightings") %>%
  distinct(scenario, year, value) %>%
  mutate(value = round(value*100, 2)) %>%
  group_by(scenario) %>%
  mutate(avg = mean(value)) %>%
  spread(year, value) %>%
  relocate(avg, .after = `2018`)

head(good_out_table)
```

## Summarize by Other Areas

```{r}
# Determine which other areas each observation would fall within
good_where_be_whales_outside <- st_join(good_outside_whales, other_areas_sf)

# Summarize by year and reserve design
good_summary_by_year_outside <- good_where_be_whales_outside %>%
  mutate(name = ifelse(is.na(name), "Other areas", name)) %>%
  st_drop_geometry() %>%
  group_by(year, name) %>%
  summarize(n_sightings = n_distinct(obs),
            n_adults = sum(adults, na.rm = T),
            n_calves = sum(calves, na.rm = T),
            n_whales = sum(total, na.rm = T)) %>%
  ungroup() %>%
  left_join(good_totals_by_year, by = "year") %>%
  mutate(prop_sightings = n_sightings/n_sightings_total,
         prop_adults = n_adults/n_adults_total,
         prop_calves = n_calves/n_calves_total,
         prop_whales = n_whales/n_whales_total) %>%
  dplyr::select(year, name, contains("prop")) %>%
  gather("obs_type", "value", contains("prop")) %>%
  mutate(obs_type = str_replace(obs_type, "prop_", "")) 
```

```{r}
# Make summary table
good_other_table <- good_summary_by_year_outside %>%
  dplyr::filter(obs_type == "sightings") %>%
  distinct(name, year, value) %>%
  mutate(value = round(value*100, 2)) %>%
  group_by(name) %>%
  mutate(avg = round(mean(value), 2)) %>%
  spread(year, value) %>%
  relocate(avg, .after = `2018`)

head(good_other_table)
```
