==> scripts/03_gfw_anchorages_and_voyages.Rmd <==
---
title: "Dominica: Identify anchorages in Dominica and voyages starting or ending there"
output: html_document
date: "2023-05-01"
---

```{r setup, include=FALSE}
# Packages
library(DBI)
library(bigrquery)
library(networkD3) # Make flow diagram
library(webshot) # Save static version of flow diagram

# Load common helper script
source(here::here("common.R"))

# Load reserve boundaries
load(file.path(prj_cachalote_dir, "data", "processed", "reserve_boundaries.Rdata"))

# Load Dominica boundaries
load(file.path(prj_cachalote_dir, "data", "processed", "dominica_boundaries.Rdata"))
```

# Introduction

This script queries the GFW port and voyages tables to find trips that start/end in Dominica. 

# Identify ports/anchorages in Dominica

```{r}
sql_port_index <- "#StandardSQL

  # Select Anchorage information for all ports in Dominica
  
  SELECT 
    s2id AS anchorage_id,
    label AS anchorage_name,
    sublabel AS anchorage_sublabel,
    lat,
    lon,
    total_visits,
    top_destination,
    iso3,
    distance_from_shore_m,
    dock
  FROM 
    `world-fishing-827.anchorages.named_anchorages_v20230302`
  WHERE 
    iso3 = 'DMA'
  ORDER BY anchorage_name"
```

```{r query-port-index, eval=F}
# Table to save output
port_index_table <- bq_table(project = bq_project,
                             table = "port_index",
                             dataset = "dominica")

if(bq_table_exists(port_index_table)){
  bq_table_delete(port_index_table)
}else{
  bq_table_create(port_index_table)
}

# Run query - We have 25 named anchorages 
dom_ports <- bq_project_query(bq_project, 
                              query = sql_port_index,
                              destination_table = port_index_table,
                              use_legacy_sql = FALSE,
                              allowLargeResults = TRUE) %>%
  bq_table_download()


```

```{r}
# Make a couple of small name adjustments
dom_ports_out <- dom_ports %>%
  mutate(anchorage_name_revised = ifelse(anchorage_name == "DMA-4", anchorage_sublabel, anchorage_name),
         port_group = case_when(anchorage_name == "PORTSMOUTH" ~ "Portsmouth",
                                anchorage_name %in% c("ROSEAU", "POINTE MICHEL") ~ "Roseau/Pointe Michel",
                                anchorage_name %in% c("DMA-4", "SAINT JOSEPH") ~ "Salisbury/Saint Joseph"))

# Save locally too
write_csv(dom_ports_out, file.path(prj_cachalote_dir, "data", "processed", "dominica_anchorages_gfw.csv"))
```


```{r}
# Make ports spatial
dom_ports_sf <- dom_ports_out %>%
  st_as_sf(coords = c("lon", "lat"),
           crs = 4326,
           remove = F) %>%
  sf::st_transform(crs = st_crs(prj_moll))  # convert points to Mollewide

# Plot 
anchorages_plot <- ggplot()+
  geom_sf(data = dominica_sf_moll, color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = reserve_boundaries_sf, aes(fill = fct_rev(as.factor(scenario))), alpha = 1, color = NA)+
  geom_sf(data = dom_ports_sf, aes(color = anchorage_name_revised), alpha = 1, size = 0.5)+
  coord_sf(xlim = plot_bbox$xlim, ylim = plot_bbox$ylim)+
  theme_basic()+
  guides(color = guide_legend(title = "Name", nrow = 3),
         fill = "none")+
  labs(title = "Named anchorages in Dominica identified by GFW")

ggsave(file.path(prj_cachalote_dir, "figures", "dominica_anchorages.png"), anchorages_plot, height = 8, width = 8)

anchorages_plot
```

```{r}
# Now calculate centroid by group
dom_ports_centroid_sf <- dom_ports_sf %>%
  group_by(port_group) %>%
  mutate(lat_cen = mean(lat),
         lon_cen = mean(lon)) %>%
  ungroup() %>%
  st_drop_geometry() %>%
  st_as_sf(coords = c("lon_cen", "lat_cen"),
           crs = 4326,
           remove = F) %>%
  sf::st_transform(crs = st_crs(prj_moll)) # convert points to Mollewide

# Plot 
anchorage_centroid_plot <- ggplot()+
  geom_sf(data = dominica_sf_moll, color = "black", fill = "grey", size = 0.2)+
  geom_sf(data = reserve_boundaries_sf, aes(fill = fct_rev(as.factor(scenario))), alpha = 1, color = NA)+
  geom_sf(data = dom_ports_centroid_sf, aes(shape = port_group))+
  coord_sf(xlim = plot_bbox$xlim, ylim = plot_bbox$ylim)+
  theme_basic()+
  guides(color = guide_legend(title = "Anchorage group", nrow = 3),
         fill = "none")+
  labs(title = "Anchorage groups in Dominica identified by GFW")

ggsave(file.path(prj_cachalote_dir, "figures", "dominica_anchorage_groups.png"), anchorage_centroid_plot, height = 8, width = 8)

anchorage_centroid_plot
```

```{r}
# Save anchorage information
save(dom_ports_centroid_sf, 
     file = file.path(prj_cachalote_dir, "data", "processed", "dominica_anchorage_centroids_gfw.Rdata"))
```

# Identify voyages starting or ending at anchorages in Dominica

```{r}
sql_dom_voyages <- "WITH 

  #########################
  # Anchorage info we want to add to this table

  anchorages AS (
      SELECT
        s2id AS anchorage_id,
        label AS anchorage_name,
        sublabel AS anchorage_sublabel,
        iso3
      FROM 
        `world-fishing-827.anchorages.named_anchorages_v20230302`
  ),

  #########################
  # List of known bad mmsi numbers

  bad_mmsi AS (
      SELECT
        ssvid
      FROM
        `world-fishing-827.gfw_research.bad_mmsi`
      CROSS JOIN
      UNNEST(ssvid) AS ssvid
  ),
  
  #########################
  # All voyages by vessels with a port visit confidence 4 - if we want to lower this confidence, we'll need to pull from a different table
  
  good_voyages AS(
      SELECT 
        *,
        EXTRACT(YEAR FROM trip_start) AS start_year,
        EXTRACT(YEAR FROM trip_end) AS end_year
      FROM  
        `world-fishing-827.pipe_production_v20201001.proto_voyages_c4`
      WHERE
        EXTRACT(YEAR FROM trip_start) > 2016 
        AND EXTRACT(YEAR FROM trip_end) < 2023
        AND CAST(ssvid AS int64) NOT IN (
                  SELECT
                    ssvid
                  FROM
                    bad_mmsi)
          AND trip_start_anchorage_id IN (
                  SELECT
                    anchorage_id
                  FROM
                    `emlab-gcp.dominica.port_index`)
      UNION ALL
      SELECT 
        *,
        EXTRACT(YEAR FROM trip_start) AS start_year,
        EXTRACT(YEAR FROM trip_end) AS end_year
      FROM  
        `world-fishing-827.pipe_production_v20201001.proto_voyages_c4`
      WHERE
        EXTRACT(YEAR FROM trip_start) > 2016 
        AND EXTRACT(YEAR FROM trip_end) < 2023
        AND CAST(ssvid AS int64) NOT IN (
                  SELECT
                    ssvid
                  FROM
                    bad_mmsi)
          AND trip_end_anchorage_id IN (
                  SELECT
                    anchorage_id
                  FROM
                    `emlab-gcp.dominica.port_index`)
          AND trip_start_anchorage_id NOT IN (
                  SELECT
                    anchorage_id
                  FROM
                    `emlab-gcp.dominica.port_index`)
      )


  SELECT 
    ssvid,
    trip_id,
    trip_start AS start_timestamp,
    start_year,
    trip_start_visit_id AS start_visit_id,
    g.trip_start_anchorage_id as start_anchorage_id,
    c.anchorage_name AS start_anchorage_name,
    c.anchorage_sublabel AS start_anchorage_sublabel,
    c.iso3 AS start_iso3,
    IF(c.iso3 = 'DMA', TRUE, FALSE) AS start_is_DMA,
    trip_end AS end_timestamp,
    end_year,
    g.trip_end_anchorage_id as end_anchorage_id,
    d.anchorage_name AS end_anchorage_name,
    d.anchorage_sublabel AS end_anchorage_sublabel,
    d.iso3 AS end_iso3,
    IF(d.iso3 = 'DMA', TRUE, FALSE) AS end_is_DMA,
    DATETIME_DIFF(trip_end, trip_start, MINUTE)/60 AS voyage_duration_h
  FROM good_voyages as g
  INNER JOIN anchorages as c
  ON g.trip_start_anchorage_id = c.anchorage_id
  INNER JOIN anchorages as d
  ON g.trip_end_anchorage_id = d.anchorage_id"
```

```{r query-voyages, eval=F}
# Table to save output
voyage_table <- bq_table(project = bq_project,
                         table = "dominica_voyages_2017_2022_raw",
                         dataset = "dominica")

if(bq_table_exists(voyage_table)){
  bq_table_delete(voyage_table)
}else{
  bq_table_create(voyage_table)
}

# Run query
dom_voyages <- bq_project_query(bq_project, 
                                query = sql_dom_voyages,
                                destination_table = voyage_table,
                                use_legacy_sql = FALSE,
                                allowLargeResults = TRUE) %>%
  bq_table_download()

# Save locally too
write_csv(dom_voyages, file.path(prj_cachalote_dir, "data", "processed", "dominica_voyages_gfw_2017_2022_raw.csv"))
```

# Start/end anchorages outside of Dominica

During the process of getting our preliminary results, we determined that some of the ISO 3 codes associated with start/end ports may be wrong. Therefore, let's look more closely at these other ports and correct where necessary. 

```{r}
other_anchorage_sql <- 'WITH

  ports as (
    SELECT 
      *,
      "origin" as type
    FROM 
      `world-fishing-827.anchorages.named_anchorages_v20230302`
    WHERE 
      s2id IN (
        SELECT DISTINCT
          start_anchorage_id
        FROM
          `emlab-gcp.dominica.dominica_voyages_2017_2022_raw`
        WHERE
          start_is_DMA = FALSE
      )
    UNION ALL
    SELECT 
      *,
      "destination" as type
    FROM 
    `world-fishing-827.anchorages.named_anchorages_v20230302`
    WHERE 
      s2id IN (
        SELECT DISTINCT
          end_anchorage_id
        FROM
          `emlab-gcp.dominica.dominica_voyages_2017_2022_raw`
        WHERE
          end_is_DMA = FALSE
      )
  )

  SELECT DISTINCT
    type,
    s2id AS anchorage_id,
    label AS anchorage_name,
    sublabel AS anchorage_sublabel,
    lat,
    lon,
    total_visits,
    top_destination,
    iso3,
    distance_from_shore_m,
    dock
  FROM 
    ports 
  ORDER BY
    iso3,
    anchorage_name'
```

```{r query-voyages, eval=F}
# Table to save output
other_anchorage_table <- bq_table(project = bq_project,
                            table = "origin_and_destination_ports",
                            dataset = "dominica")

if(bq_table_exists(other_anchorage_table)){
  bq_table_delete(other_anchorage_table)
}else{
  bq_table_create(other_anchorage_table)
}

# Run query
other_ports <- bq_project_query(bq_project, 
                                query = other_anchorage_sql,
                                destination_table = other_anchorage_table,
                                use_legacy_sql = FALSE,
                                allowLargeResults = TRUE) %>%
  bq_table_download()
```
Classify by origin or destination port (or both).

```{r}
# Classify
other_ports_sf <- other_ports %>%
  group_by(anchorage_id) %>%
  mutate(category = case_when(length(type) == 2 ~ "Both",
                              type == "origin" ~ "Origin",
                              type == "destination" ~ "Destination"),
         category = factor(category, levels = c("Origin", "Destination", "Both"))) %>%
  distinct(anchorage_id, anchorage_name, lat, lon, iso3, category) %>%
  st_as_sf(coords = c("lon", "lat"),
           crs = 4326,
           remove = F) %>%
  sf::st_transform(crs = st_crs(prj_moll)) # convert points to Mollewide

# Plot 
other_port_plot <- ggplot()+
  geom_sf(data = other_ports_sf, aes(color = category), size = 0.5)+
  theme_basic()+
  map_layer_moll+
  guides(color = guide_legend(title = "Anchorage group", nrow = 3),
         fill = "none")+
  labs(title = "Start and end ports for voyages to Dominica as identified by GFW")

ggsave(file.path(prj_cachalote_dir, "figures", "dominica_start_and_end_ports.png"), other_port_plot, height = 8, width = 8)

other_port_plot
```

## Correct mislabeled ISO 3 codes

We've noticed that the distances traveled for certain start/end ports based on ISO 3 code don't seem right. Particularly the following: 
- FRA
- NLD

Let's look more closely at these and relabel them if necessary. Starting with "FRA"... We've definitely got a few ports in the Caribbean that should be assigned the ISO 3 codes of their respective countries or territories. 

```{r}
# Get french ports
fra_ports <- other_ports_sf %>%
  dplyr::filter(iso3 == "FRA")

# Take a look - We've definitely got three countries in the Caribbean that shouldn't be labeled as "FRA"
fra_port_plot <- ggplot()+
  geom_sf(data = fra_ports, aes(color = category), size = 1)+
  theme_basic()+
  map_layer_moll+
  guides(color = guide_legend(title = "Anchorage group", nrow = 3),
         fill = "none")

fra_port_plot
```

```{r}
# Fix the mislabeled ports
# LE GOSIER, DESHAIES, BAIE DU MARIGOT are in Guadeloupe
# SCHOELCHER is in Martineque; MTQ-16 is probably in Martinique
# ST MARTIN is in Saint Martin; MAF-2 is probably in Saint Martin
fra_ports_fix <- fra_ports %>%
  dplyr::filter(lat < 20 & lon < -50) %>%
  mutate(iso3_fixed = case_when(anchorage_name %in% c("BAIE DU MARIGOT", "DESHAIES", "LE GOSIER") ~ "GLP",
                              anchorage_name %in% c("MAF-2", "ST MARTIN") ~ "MAF",
                              anchorage_name %in% c("MTQ-16", "SCHOELCHER") ~ "MTQ")) %>%
  st_drop_geometry()
```

And repeat the process for "NLD"... Same deal, there are some overseas territories of the Netherlands that should have been assigned the ISO 3 codes of the administering territory. 

```{r}
# Get french ports
nld_ports <- other_ports_sf %>%
  dplyr::filter(iso3 == "NLD")

# Take a look - We've definitely got three countries in the Caribbean that shouldn't be labeled as "FRA"
nld_port_plot <- ggplot()+
  geom_sf(data = nld_ports, aes(color = category), size = 1)+
  theme_basic()+
  map_layer_moll+
  guides(color = guide_legend(title = "Anchorage", nrow = 3),
         fill = "none")

nld_port_plot
```

```{r}
# Fix the mislabeled ports
# ST MAARTEN is in Saint Maarten
# The rest look to be in the Netherlands
nld_ports_fix <- nld_ports %>%
  dplyr::filter(lat < 20 & lon < -50) %>%
  mutate(iso3_fixed = case_when(anchorage_name == c("ST MAARTEN") ~ "SXM",
                                T ~ "NLD")) %>%
  st_drop_geometry()
```

```{r}
# Combine
ports_fix <- fra_ports_fix %>%
  bind_rows(nld_ports_fix)

# Now add these back to the port lookup file
other_ports_out <- other_ports_sf %>%
  left_join(ports_fix %>% dplyr::select(anchorage_id, iso3_fixed), by = "anchorage_id") %>%
  mutate(iso3_fixed = ifelse(!is.na(iso3_fixed), iso3_fixed, iso3))

# Save locally too
save(other_ports_out, file = file.path(prj_cachalote_dir, "data", "processed", "origin_and_destination_ports.Rdata"))
```

## Add port corrections to voyage table

```{r}
# Correct ports
dom_voyages_fixed <- dom_voyages %>%
  left_join(ports_fix %>% dplyr::select(anchorage_id, start_iso3_fixed = iso3_fixed), by = c("start_anchorage_id" = "anchorage_id")) %>%
  left_join(ports_fix %>% dplyr::select(anchorage_id, end_iso3_fixed = iso3_fixed), by = c("end_anchorage_id" = "anchorage_id")) %>%
  mutate(start_iso3 = ifelse(!is.na(start_iso3_fixed), start_iso3_fixed, start_iso3),
         end_iso3 = ifelse(!is.na(end_iso3_fixed), end_iso3_fixed, end_iso3)) %>%
  dplyr::select(-start_iso3_fixed, -end_iso3_fixed)
```

## Add more vessel details

Before we upload the corrected voyage table to Bigquery, let's also add in some additional vessel classifications from the national AIS data where possible. 

```{r}
# Load file of domestic AIS data
load(file = file.path(prj_cachalote_dir, "data", "processed", "ais", "Dominica_AIS_2012_2019_processed.Rdata")) 

# Get vessel types and categories
ais_vessel_info <- ais_dat_dist %>%
  st_drop_geometry() %>%
  distinct(mmsi, category) %>%
  mutate(ssvid = as.character(mmsi)) %>%
  dplyr::select(-mmsi)

# Figure out which have more than one category assigned to a mmsi number
n_categories <- ais_vessel_info %>%
  group_by(ssvid) %>%
  summarize(n = n(),
            categories = paste0(category, collapse = ", "))

# We just have two that have multiple categories. We'll make them merchant vessels. 
mult_categories <- n_categories %>%
  dplyr::filter(n > 1)

# Create output vessel type table
ais_vessel_info_keep <- ais_vessel_info %>%
  mutate(category = ifelse(ssvid %in% mult_categories$ssvid, "merchant", category)) %>%
  distinct(ssvid, category)
```

```{r}
# Add to vessel table
dom_voyages_final <- dom_voyages_fixed %>%
  left_join(ais_vessel_info_keep, by = c("ssvid"))
```

## Upload corrected voyage table

```{r}
# Table to save output
voyage_table_fixed <- bq_table(project = bq_project,
                               table = "dominica_voyages_2017_2022_fixed",
                               dataset = "dominica")

if(bq_table_exists(voyage_table_fixed)){
  bq_table_delete(voyage_table_fixed)
}else{
  bq_table_create(voyage_table_fixed)
}

# Upload
voyage_table_fixed %>%
  bq_table_upload(fields = dom_voyages_final, values = dom_voyages_final,
                  write_disposition = 'WRITE_TRUNCATE')

# Save locally too
write_csv(dom_voyages_final, file.path(prj_cachalote_dir, "data", "processed", "dominica_voyages_gfw_2017_2022_fixed.csv"))
```

# Summarize

First, lets figure out how many inbound, outgoing, and domestic voyages we have between 2017 and 2022. Then let's make sure we don't have any voyages that don't meet one of these three classifications.  

## Overall trip types

```{r}
# 2,068 domestic voyages (starting and ending in Dominica)
domestic_voyages <- dom_voyages_final %>%
  dplyr::filter(start_is_DMA & end_is_DMA) %>%
  distinct(trip_id)

# 10,837 inbound voyages
in_voyages <- dom_voyages_final %>%
  dplyr::filter(!start_is_DMA & end_is_DMA) %>%
  distinct(trip_id)

# 11,078 outgoing voyages
out_voyages <- dom_voyages_final %>%
  dplyr::filter(start_is_DMA & !end_is_DMA) %>%
  distinct(trip_id)

# 23,983 total voyages
all_voyages <- c(domestic_voyages$trip_id, in_voyages$trip_id, out_voyages$trip_id)

# Are we missing any? No. 
other_voyages <- dom_voyages_final %>%
  dplyr::filter(!(trip_id %in% all_voyages)) %>%
  distinct(trip_id)

# Do we have duplicate rows in our voyage table? No. 
n_rows_per_trip <- dom_voyages_final %>%
  group_by(trip_id, ssvid) %>%
  summarize(n_rows = n()) %>%
  dplyr::filter(n_rows > 1)
```

```{r}
# Let's do some simple classifications for our voyages: 
voyage_classification <- dom_voyages_final %>%
  mutate(classification = case_when(trip_id %in% domestic_voyages$trip_id ~ "Domestic",
                                    trip_id %in% in_voyages$trip_id ~ "Inbound",
                                    trip_id %in% out_voyages$trip_id ~ "Outbound")) %>%
  group_by(classification) %>%
  summarize(n_voyages = n(),
            average_duration = mean(voyage_duration_h),
            total_duration = sum(voyage_duration_h),
            min_duration = min(voyage_duration_h),
            max_duration = max(voyage_duration_h))

head(voyage_classification)
```

## Frequency of voyages by start/end country

```{r}
# Summarize by start/end - country level only
freq_voyages_by_iso3 <- dom_voyages_final %>%
 mutate(classification = case_when(trip_id %in% domestic_voyages$trip_id ~ "Domestic",
                                    trip_id %in% in_voyages$trip_id ~ "Inbound",
                                    trip_id %in% out_voyages$trip_id ~ "Outbound")) %>%
  group_by(classification, start_iso3, end_iso3) %>%
  summarize(n_voyages = n_distinct(trip_id),
            avg_duration_h = mean(voyage_duration_h))
```


```{r}
# Summarize by start/end - country level only
freq_voyages_by_iso3_and_year <- dom_voyages_final %>%
  mutate(classification = case_when(trip_id %in% domestic_voyages$trip_id ~ "Domestic",
                                    trip_id %in% in_voyages$trip_id ~ "Inbound",
                                    trip_id %in% out_voyages$trip_id ~ "Outbound")) %>%
  group_by(start_year, classification, start_iso3, end_iso3) %>%
  summarize(n_voyages = n_distinct(trip_id),
            avg_duration_h = mean(voyage_duration_h))
```

```{r}
# Plot flow diagram of voyages to Dominica
links_in <- freq_voyages_by_iso3_and_year %>%
  dplyr::filter(start_year == 2022) %>%
  ungroup() %>%
  dplyr::filter(start_iso3 != "DMA" & end_iso3 == "DMA") %>%
  dplyr::select(source = start_iso3, target = end_iso3, value = n_voyages)
links_in$source <- paste(links_in$source, " ", sep="")

# Plot flow diagram of total landings made in each country that go toward each destination
links_out <- freq_voyages_by_iso3_and_year %>%
  dplyr::filter(start_year == 2022) %>%
  ungroup() %>%
  dplyr::filter(start_iso3 == "DMA") %>%
  dplyr::select(source = start_iso3, target = end_iso3, value = n_voyages)

# Join to previous df
links_all <- links_in %>%
  bind_rows(links_out)

nodes <- tibble(name=c(as.character(links_all$source), as.character(links_all$target)) %>% unique()) %>%
  mutate(region = countrycode(str_replace(name, " ", ""), "iso3c", "region"),
         group = case_when(region == "Latin America & Caribbean" ~ "group_1",
                           region == "Sub-Saharan Africa" ~ "group_2",
                           region == "Europe & Central Asia" ~ "group_3",
                           region == "North America" ~ "group_4",
                           T ~ "group_5"),
         group = as.factor(group)) %>%
  as.data.frame()

# With networkD3, connection must be provided using id, not using real name like in the links dataframe.. So we need to reformat it.
links_all$IDsource=match(links_all$source, nodes$name)-1 
links_all$IDtarget=match(links_all$target, nodes$name)-1

# Color scale
# rev(RColorBrewer::brewer.pal(3, "Set2"))
my_color <- 'd3.scaleOrdinal() .domain(["group_1", "group_2", "group_3", "group_4", "group_5"]) .range(["#8DA0CB", "#FC8D62", "#66C2A5", "#6ca992", "grey"])'

# Make plot
sn <- sankeyNetwork(Links = links_all, Nodes = nodes,
              Source = "IDsource", Target = "IDtarget",
              Value = "value", NodeID = "name",
              sinksRight=F, nodeWidth=40, fontSize=13, nodePadding=20,
              iterations = 100,
              units = "Voyages",
              NodeGroup ="group",
              colourScale = my_color)

# Save
saveNetwork(sn, file.path(prj_cachalote_dir, "figures", "voyages_in_and_out_of_dominica_by_iso3.html"))

# # And convert to a static PNG version
# webshot(here("results", "voyages_in_and_out_of_dominica_by_iso3.html"),
#         here("results", "voyages_in_and_out_of_dominica_by_iso3.png"), vwidth = 2500, vheight = 1500)

sn
```

```{r}
# Summarize by start/end - by port are within Dominica
freq_voyages_by_iso3_and_port <- dom_voyages_final %>%
  mutate(start_anchor_group = case_when(start_anchorage_name %in% c("POINTE MICHEL", "ROSEAU") ~ "Roseau/Pointe Michel",
                                        start_anchorage_name == "PORTSMOUTH" ~ "Portsmouth",
                                        start_anchorage_name %in% c("DMA-4", "SAINT JOSEPH") ~ "Saint Joseph/Salisbury"),
         end_anchor_group = case_when(end_anchorage_name %in% c("POINTE MICHEL", "ROSEAU") ~ "Roseau/Pointe Michel",
                                      end_anchorage_name == "PORTSMOUTH" ~ "Portsmouth",
                                      end_anchorage_name %in% c("DMA-4", "SAINT JOSEPH") ~ "Saint Joseph/Salisbury")) %>%
  mutate(start = ifelse(start_iso3 == "DMA", start_anchor_group, start_iso3),
         end = ifelse(end_iso3 == "DMA", end_anchor_group, end_iso3)) %>%
  group_by(start_year, start, end) %>%
  summarize(n_voyages = n_distinct(trip_id),
            avg_duration_h = mean(voyage_duration_h))

# Remove those with the same start/end destination
freq_voyages_by_iso3_and_port <- freq_voyages_by_iso3_and_port %>%
  dplyr::filter(start != end)
```

```{r}
# Plot flow diagram of voyages to Dominica
links_in_p <- freq_voyages_by_iso3_and_port %>%
  dplyr::filter(start_year == 2022) %>%
  ungroup() %>%
  dplyr::filter(end %in% c("Roseau/Pointe Michel", "Saint Joseph/Salisbury", "Portsmouth")) %>%
  dplyr::select(source = start, target = end, value = n_voyages)
links_in_p$source <- paste(links_in_p$source, " ", sep="")

# Plot flow diagram of total landings made in each country that go toward each destination
links_out_p <- freq_voyages_by_iso3_and_port %>%
  dplyr::filter(start_year == 2022) %>%
  ungroup() %>%
  dplyr::filter(start %in% c("Roseau/Pointe Michel", "Saint Joseph/Salisbury", "Portsmouth")) %>%
  dplyr::select(source = start, target = end, value = n_voyages)
links_out_p$target <- paste(links_out_p$target, "  ", sep="")

# Join to previous df
links_all_p <- links_in_p %>%
  bind_rows(links_out_p)

nodes_p <- tibble(name=c(as.character(links_all_p$source), as.character(links_all_p$target)) %>% unique()) %>%
  mutate(region = countrycode(str_replace_all(name, " ", ""), "iso3c", "region"),
         group = case_when(region == "Latin America & Caribbean" ~ "group_1",
                           region == "Sub-Saharan Africa" ~ "group_2",
                           region == "Europe & Central Asia" ~ "group_3",
                           region == "North America" ~ "group_4",
                           T ~ "group_5"),
         group = as.factor(group)) %>%
  as.data.frame()

# With networkD3, connection must be provided using id, not using real name like in the links dataframe.. So we need to reformat it.
links_all_p$IDsource=match(links_all_p$source, nodes_p$name)-1 
links_all_p$IDtarget=match(links_all_p$target, nodes_p$name)-1

# Color scale
# rev(RColorBrewer::brewer.pal(3, "Set2"))
my_color <- 'd3.scaleOrdinal() .domain(["group_1", "group_2", "group_3", "group_4", "group_5"]) .range(["#8DA0CB", "#FC8D62", "#66C2A5", "#6ca992", "grey"])'

# Make plot
sn_p <- sankeyNetwork(Links = links_all_p, Nodes = nodes_p,
              Source = "IDsource", Target = "IDtarget",
              Value = "value", NodeID = "name",
              sinksRight=F, nodeWidth=40, fontSize=13, nodePadding=20,
              iterations = 500,
              units = "Voyages",
              NodeGroup ="group",
              colourScale = my_color)

# Save
saveNetwork(sn_p, file.path(prj_cachalote_dir, "figures", "voyages_in_and_out_of_dominica_by_iso3_and_port.html"))

# # And convert to a static PNG version
# webshot(here("results", "voyages_in_and_out_of_dominica_by_iso3.html"),
#         here("results", "voyages_in_and_out_of_dominica_by_iso3.png"), vwidth = 2500, vheight = 1500)

sn_p
```
