==> scripts/05_gfw_diagnostics.Rmd <==
---
title: "Dominica: Diagnostics"
output: html_document
date: "2023-05-08"
---

```{r setup, include=FALSE}
# Packages
library(geosphere)
library(ggridges)

# Load common helper script
source(here::here("common.R"))

# Load reserve boundaries
load(here::here("results", "reserve_boundaries.Rdata"))

# Load voyage data we need
dom_voyages <- read_csv(file.path(prj_cachalote_dir, "data", "processed", "dominica_voyages_gfw_2017_2022.csv")) # All voyages - 23,983
trips_in <- read_csv(file.path(prj_cachalote_dir, "data", "processed", "trips_in_summary_table_v2.csv"))
trips_out <- read_csv(file.path(prj_cachalote_dir, "data", "processed", "trips_out_summary_table_v2.csv"))

# Get anchorage centroids 
load(file.path(prj_cachalote_dir, "data", "processed", "dominica_anchorage_centroids_gfw.Rdata"))
centroids <- dom_ports_centroid_sf %>%
  sf::st_transform(crs = prj_moll)
```

# Check for missing values

Let's figure out whether we're missing data for any voyages and whether any trips should be excluded.  

## Inbound voyages

```{r}
# Inbound voyages
inbound_voyages_pos <- trips_in %>% distinct(trip_id) %>% pull()
inbound_voyages_all <- dom_voyages %>% dplyr::filter(!start_is_DMA & end_is_DMA) %>% distinct(trip_id) %>% pull()

# Do we have the same trips represented in the data? Yes. 
abs(length(inbound_voyages_pos) - length(inbound_voyages_all))

# Do we have any trips with missing positional information. Yes - 7
inbound_voyages_nas <- trips_in %>% dplyr::filter(is.na(lat_start) | is.na(lon_start) | is.na(lat_end) | is.na(lon_end)) %>%
  distinct(trip_id) %>% pull()

# Good inbound trips
inbound_voyages_good <- trips_in %>% 
  dplyr::filter(!is.na(lat_start) & !is.na(lon_start) & !is.na(lat_end) & !is.na(lon_end)) %>%
  dplyr::filter(!is.na(scenario)) %>%
  distinct(trip_id) %>% pull()

# Determine the proportion of trips that would be affected by each reserve scenario
inbound_prop_affected <- trips_in %>%
  dplyr::filter(trip_id %in% inbound_voyages_good) %>%
  summarize(prop_unaffected = length(unique(trip_id[is.na(scenario)]))/length(inbound_voyages_good),
            prop_affected_scenario_1 = length(unique(trip_id[scenario == 1]))/length(inbound_voyages_good),
            prop_affected_scenario_2 = length(unique(trip_id[scenario == 2]))/length(inbound_voyages_good),
            prop_affected_scenario_3 = length(unique(trip_id[scenario == 3]))/length(inbound_voyages_good)) %>%
  gather()

head(inbound_prop_affected)
```

## Outbound voyages

```{r}
# Inbound voyages
outbound_voyages_pos <- trips_out %>% distinct(trip_id) %>% pull()
outbound_voyages_all <- dom_voyages %>% dplyr::filter(start_is_DMA & !end_is_DMA) %>% distinct(trip_id) %>% pull()

# Do we have the same trips represented in the data? Yes. 
abs(length(outbound_voyages_pos) - length(outbound_voyages_all))

# Do we have any trips with missing positional information. Yes - 5
outbound_voyages_nas <- trips_out %>% dplyr::filter(is.na(lat_start) | is.na(lon_start) | is.na(lat_end) | is.na(lon_end)) %>%
  distinct(trip_id) %>% pull()

# Good outbound trips
outbound_voyages_good <- trips_out %>% 
  dplyr::filter(!is.na(lat_start) & !is.na(lon_start) & !is.na(lat_end) & !is.na(lon_end)) %>%
  dplyr::filter(!is.na(scenario)) %>%
  distinct(trip_id) %>% pull()

# Determine the proportion of trips that would be affected by each reserve scenario
outbound_prop_affected <- trips_out %>%
  dplyr::filter(trip_id %in% outbound_voyages_good) %>%
  summarize(prop_unaffected = length(unique(trip_id[is.na(scenario)]))/length(outbound_voyages_good),
            prop_affected_scenario_1 = length(unique(trip_id[scenario == 1]))/length(outbound_voyages_good),
            prop_affected_scenario_2 = length(unique(trip_id[scenario == 2]))/length(outbound_voyages_good),
            prop_affected_scenario_3 = length(unique(trip_id[scenario == 3]))/length(outbound_voyages_good)) %>%
  gather()

head(outbound_prop_affected)
```

# Distance of start/end positions to anchorage coordinates and grouped centroids

Since we're using the centroid anchorage coordinates for out distance matrices, we might need to add a bit of extra time/distance based on the actual end points of each trip. Let's see how far our start/end coordinates are from the anchorage coordinates and centroids. 

We'll then filter for trips that 

## Inbound trips

```{r}
# Calculate distance for our good trips
trips_in_dist <- trips_in %>%
  dplyr::filter(trip_id %in% inbound_voyages_good) %>%
  dplyr::distinct(trip_id, vessel_type, lat_end, lon_end, end_anchorage_id) %>%
  left_join(centroids %>% st_drop_geometry() %>% dplyr::select(anchorage_id, lat, lon, lat_cen, lon_cen), by = c("end_anchorage_id" = "anchorage_id")) %>%
  rowwise() %>%
  mutate(dist_from_port = distm(c(lon, lat), c(lon_end, lat_end), fun=distHaversine)/1000,
         dist_from_port_centroid = distm(c(lon_cen, lat_cen), c(lon_end, lat_end), fun=distHaversine)/1000)

# Get bad trips
bad_trips_in <- trips_in_dist %>%
  dplyr::filter(dist_from_port > 5) %>%
  pull(trip_id)

# Percentage of trips being excluded
percent_bad_in <- length(bad_trips_in)/length(inbound_voyages_good)
trips_remaining_in <- length(inbound_voyages_good) - length(bad_trips_in)
```

```{r}
# Make 
trips_in_dist_plot <- trips_in_dist %>%
  dplyr::filter(!trip_id %in% bad_trips_in) %>%
  ggplot()+
  geom_density_ridges(aes(x = dist_from_port_centroid, y = vessel_type), alpha = 0.5, fill = "blue")+
  geom_density_ridges(aes(x = dist_from_port, y = vessel_type), alpha = 0.5, fill = "red")+
  theme_basic()+
  labs(x = "Distance between last AIS position and anchorage (km, red) or anchorage centroid (km, blue)", y = "Vessel type", 
       title = paste0("Inbound trips: Last AIS position relative to anchorage coordinates (N = ",  format(trips_remaining_in, big.mark = ","), ")"))

ggsave(file.path(prj_cachalote_dir, "figures", "inbound_trips_end_dist_from_port.png"), trips_in_dist_plot, height = 6, width = 7.5)

trips_in_dist_plot
```

## Outbound trips

```{r}
trips_out_dist <- trips_out %>%
  dplyr::filter(trip_id %in% outbound_voyages_good) %>%
  dplyr::distinct(trip_id, vessel_type, lat_start, lon_start, start_anchorage_id) %>%
  left_join(centroids %>% st_drop_geometry() %>% dplyr::select(anchorage_id, lat, lon, lat_cen, lon_cen), by = c("start_anchorage_id" = "anchorage_id")) %>%
  rowwise() %>%
  mutate(dist_from_port = distm(c(lon, lat), c(lon_start, lat_start), fun=distHaversine)/1000,
         dist_from_port_centroid = distm(c(lon_cen, lat_cen), c(lon_start, lat_start), fun=distHaversine)/1000)

# Get bad trips
bad_trips_out <- trips_out_dist %>%
  dplyr::filter(dist_from_port > 10) %>%
  pull(trip_id)

# Percentage of trips being excluded
percent_bad_out <- length(bad_trips_out)/length(outbound_voyages_good)
trips_remaining_out <- length(outbound_voyages_good) - length(bad_trips_out)
```

```{r}
# Make 
trips_out_dist_plot <- trips_out_dist %>%
  dplyr::filter(!(trip_id %in% bad_trips_out)) %>%
  ggplot()+
  geom_density_ridges(aes(x = dist_from_port_centroid, y = vessel_type), alpha = 0.5, fill = "blue")+
  geom_density_ridges(aes(x = dist_from_port, y = vessel_type), alpha = 0.5, fill = "red")+
  theme_basic()+
  labs(x = "Distance between first AIS position and anchorage (km, red) or anchorage centroid (km, blue)", y = "Vessel type", 
       title = paste0("Outbound trips: First AIS position relative to anchorage coordinates (N = ",  format(trips_remaining_out, big.mark = ","), ")"))

ggsave(file.path(prj_cachalote_dir, "figures", "outbound_trips_end_dist_from_port.png"), trips_out_dist_plot, height = 6, width = 7.5)

trips_out_dist_plot
```

# Make final list of good trips

```{r}
good_trips <- c(inbound_voyages_good[!inbound_voyages_good %in% bad_trips_in],
                outbound_voyages_good[!outbound_voyages_good %in% bad_trips_out])

save(good_trips,
     file = file.path(prj_cachalote_dir, "data", "processed", "good_trips.Rdata"))
```

# Vessels with a matching category

```{r}
trips_in_category <- trips_in %>%
  dplyr::filter(trip_id %in% good_trips) %>%
  distinct(trip_id, category) %>%
  group_by(category) %>%
  summarize(n = n()) %>%
  mutate(prop = (n/trips_remaining_in)*100)

trips_out_category <- trips_out %>%
  dplyr::filter(trip_id %in% good_trips) %>%
  distinct(trip_id, category) %>%
  group_by(category) %>%
  summarize(n = n()) %>%
  mutate(prop = (n/trips_remaining_out)*100)
```
